New paramgen module
