+++ .gitignore
@@ -1,4 +1,4 @@
-nnmnkwii/functions/_impl/_mlpg.c
+nnmnkwii/paramgen/mlpg_helper.c
 nnmnkwii/util/_linalg.c
 examples
 docs/references/generated
