+++ setup.py
@@ -38,8 +38,8 @@ ext_modules = [
         extra_compile_args=["-std=c99"],
     ),
     Extension(
-        name="nnmnkwii.functions._impl._mlpg",
-        sources=[join("nnmnkwii", "functions", "_impl", "_mlpg" + ext)],
+        name="nnmnkwii.paramgen.mlpg_helper",
+        sources=[join("nnmnkwii", "paramgen", "mlpg_helper" + ext)],
         include_dirs=[np.get_include()],
         language="c",
         extra_compile_args=["-std=c99"]
